Refactor pagination nav
Generic implementation, can now be used in all views
